Created a LogBinStats class to perform log-binomial regression.
Child class of RegressionStats, and very similar to LogitStats in
functionality, returning RR instead of OR. Built on sm.GLM(), using
a binomial distribution with a log (natural log) link function.

Partly due to my unfamiliarity with sm.GLM and partly due to intrinsic
limitations of log-binomial regression, the results are error-prone &
not optimized. `maxiter=1000` to minimize failures to converge, and a
basic set of starting values are  given. If source code of R `logbin`
package can is reviewed, perhaps this can be improved significantly. A
warning is issued upon instantiation regarding its experimental nature.